accessible/tests/mochitest/selectable/test_menulist.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<?xml-stylesheet href="../treeview.css" type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="XUL tree selectable tests">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../states.js" />
  <script type="application/javascript"
          src="../selectable.js" />

  <script type="application/javascript">
  <![CDATA[

    ////////////////////////////////////////////////////////////////////////////
    // Test

    //gA11yEventDumpID = "debug";

    var gQueue = null;

    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // menulist aka combobox

      var id = "combobox";
      var combobox = getAccessible(id);
      var comboboxList = combobox.firstChild;
      ok(isAccessible(comboboxList, [nsIAccessibleSelectable]),
         "No selectable accessible for list of " + id);

      var select = getAccessible(comboboxList, [nsIAccessibleSelectable]);
      testSelectableSelection(select, [ "cb1_item1" ]);

      select.addItemToSelection(1);
      testSelectableSelection(select, [ "cb1_item2" ], "addItemToSelection(1): ");

      select.removeItemFromSelection(1);
      testSelectableSelection(select, [ ],
                              "removeItemFromSelection(1): ");

      is(select.selectAll(), false,
         "No way to select all items in combobox '" + id + "'");
      testSelectableSelection(select, [ ], "selectAll: ");

      select.addItemToSelection(1);
      select.unselectAll();
      testSelectableSelection(select, [ ], "unselectAll: ");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
        href="https://bugzilla.mozilla.org/show_bug.cgi?id=590176"
        title="add pseudo SelectAccessible interface">
       Mozilla Bug 590176
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <menulist id="combobox">
        <menupopup>
          <menuitem id="cb1_item1" label="item1"/>
          <menuitem id="cb1_item2" label="item2"/>
        </menupopup>
      </menulist>

      <vbox id="debug"/>
    </vbox>
  </hbox>

</window>
